update: move common class to utils
